+++ changelog.md
@@ -1,3 +1,6 @@
+### [09-06-2021] RAMFS renamed to TMPFS
+* As TMPFS is better and modernish, and it was a naming typo
+
 ### [07-06-2021] MongoDB auth fixed
 * [Example of user/role creation](https://pymongo.readthedocs.io/en/stable/examples/authentication.html)
 ```